Api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Mock FastAPI for ClassManager"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.119.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.27",
    "sqlalchemy>=2.0.0",
    "cryptography>=46.0.0",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest"
]
